pyproject.toml
[project]
name = "meta-attn"
version = "0.0.1"
description = "Meta-attention for retrieval-augmented generation"
readme = "README.md"
authors = [
    { name="Alex Chen", email="azc2110@columbia.edu" },
    { name="René Sultan", email="rs4240@columbia.edu"},
]
requires-python = ">=3.10"
dependencies = [
    # sort these alphabetically
    "accelerate",
    "python-dotenv",
    "ipykernel",
    "ipywidgets",
    "matplotlib",
    "numpy",
    "openai",
    "pandas",
    "scipy",
    "tensorboard",
    "torch",
    "torchvision",
    "tqdm",
    "transformers",
    "typer",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["mattn"]

[tool.isort]
profile = "black"

[tool.pyright]
exclude = ["**/.*", "**/__pycache__", "data"]
